76 Moved MarkLogicBatchConfiguration to src/test/java
