+++ core/src/test/java/com/marklogic/spring/batch/AbstractSpringBatchTest.java
@@ -4,7 +4,6 @@ import javax.xml.bind.JAXBContext;
 import javax.xml.bind.JAXBException;
 
 import com.marklogic.client.spring.BasicConfig;
-import com.marklogic.spring.batch.configuration.MarkLogicBatchConfiguration;
 import org.springframework.batch.core.BatchStatus;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.JobExecution;
